Fix for #1726 (Deep recursion when I delete DNS record)

+++ lib/Netdot/Model/RR.pm
@@ -368,7 +368,7 @@ sub delete {
     my @cnames = RRCNAME->search(cname=>$self->get_label);
     my @mxs = RRMX->search(exchange=>$self->get_label);
     foreach my $o ( @cnames, @mxs ){
-	$o->rr->delete();
+	$o->rr->delete() if ($o->rr->get_label ne $self->get_label);
     }
     return $self->SUPER::delete();
 }
